Lots of work. Some structural changes to classes to better fit their usage.
Todo List is completed for its base functionality. More may be added to it to show progress when finishing todo items.

Events calendar nearing completion. Events can be added for any day and deleted for this day. The list view of tasks for the day populates with events for the current day. This will soon change to allow for the user to choose the date they are viewing. Last major functionality to be added is the edit button for the calendar events.

After that, methods will be implemented in the classes to give them behavior rather than just state. This will come in the form of adding contextual information to the user and adding info panels to give visuals to progress made in and for the given day.